minor code refactoring. Split schema with utililty functions
