+++ requestly-android-core/src/main/java/io/requestly/android/core/modules/hostSwitcher/HostSwitcherFragmentViewModel.kt
@@ -5,8 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.google.gson.reflect.TypeToken
 import io.requestly.android.core.KeyValueStorageManager
-import io.requestly.android.core.modules.hostSwitcher.models.HttpVerb
-import io.requestly.android.core.modules.hostSwitcher.models.Rule
+import io.requestly.android.core.modules.hostSwitcher.models.*
 
 
 typealias SwitchingRule = Rule
